+++ packages/deploy-config-sub-generator/test/headless/cf-headless.test.ts
@@ -1,9 +1,9 @@
 import hasbin from 'hasbin';
-import * as childProcess from 'child_process';
+import * as childProcess from 'node:child_process';
 import { toMatchFolder } from '@sap-ux/jest-file-matchers';
 import { readdirSync, writeFileSync } from 'node:fs';
 import { copy, existsSync } from 'fs-extra';
-import { readFile, rename } from 'fs/promises';
+import { readFile, rename } from 'node:fs/promises';
 import { rimraf } from 'rimraf';
 import { basename, join } from 'node:path';
 import CFGen from '@sap-ux/cf-deploy-config-sub-generator';