+++ packages/environment-check/test/command.test.ts
@@ -1,5 +1,5 @@
 import { spawnCommand } from '../src/command';
-import * as cp from 'child_process';
+import * as cp from 'node:child_process';
 
 jest.mock('child_process');
 const mockedCp = jest.mocked(cp, { shallow: true });