+++ packages/ui5-config/src/ui5config.ts
@@ -29,7 +29,7 @@ import { fioriToolsProxy, serveStatic } from './constants';
 import Ajv, { type ValidateFunction } from 'ajv';
 import type { SomeJSONSchema } from 'ajv/dist/types/json-schema';
 import { join, posix, relative, sep } from 'node:path';
-import { readFile } from 'fs/promises';
+import { readFile } from 'node:fs/promises';
 import yaml from 'js-yaml';
 
 /**