+++ packages/abap-deploy-config-inquirer/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @sap-ux/abap-deploy-config-inquirer
 
+## 1.6.64
+
+### Patch Changes
+
+- a9471d0: fix sonar issues
+- Updated dependencies [a9471d0]
+    - @sap-ux/axios-extension@1.25.1
+    - @sap-ux/inquirer-common@0.10.2
+    - @sap-ux/system-access@0.6.35
+    - @sap-ux/nodejs-utils@0.2.11
+    - @sap-ux/ui5-config@0.29.12
+    - @sap-ux/logger@0.7.3
+    - @sap-ux/store@1.4.1
+    - @sap-ux/fiori-generator-shared@0.13.45
+    - @sap-ux/btp-utils@1.1.6
+
+## 1.6.63
+
+### Patch Changes
+
+- @sap-ux/inquirer-common@0.10.1
+
 ## 1.6.62
 
 ### Patch Changes