+++ packages/cf-deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @sap-ux/cf-deploy-config-sub-generator
 
+## 0.2.99
+
+### Patch Changes
+
+- a9471d0: fix sonar issues
+- Updated dependencies [a9471d0]
+    - @sap-ux/cf-deploy-config-inquirer@0.5.57
+    - @sap-ux/cf-deploy-config-writer@0.3.47
+    - @sap-ux/inquirer-common@0.10.2
+    - @sap-ux/project-access@1.32.17
+    - @sap-ux/deploy-config-generator-shared@0.1.64
+    - @sap-ux/fiori-generator-shared@0.13.45
+    - @sap-ux/btp-utils@1.1.6
+
+## 0.2.98
+
+### Patch Changes
+
+- @sap-ux/inquirer-common@0.10.1
+- @sap-ux/cf-deploy-config-inquirer@0.5.56
+- @sap-ux/deploy-config-generator-shared@0.1.63
+
 ## 0.2.97
 
 ### Patch Changes