+++ packages/fe-fpm-writer/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @sap-ux/fe-fpm-writer
 
+## 0.40.1
+
+### Patch Changes
+
+- a9471d0: fix sonar issues
+- Updated dependencies [a9471d0]
+    - @sap-ux/fiori-annotation-api@0.8.1
+    - @sap-ux/project-access@1.32.17
+    - @sap-ux/logger@0.7.3
+
 ## 0.40.0
 
 ### Minor Changes