+++ packages/launch-config/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sap-ux/launch-config
 
+## 0.10.44
+
+### Patch Changes
+
+- Updated dependencies [a9471d0]
+    - @sap-ux/project-access@1.32.17
+    - @sap-ux/ui5-config@0.29.12
+    - @sap-ux/ui5-info@0.13.5
+
 ## 0.10.43
 
 ### Patch Changes