+++ packages/ui5-library-writer/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sap-ux/ui5-library-writer
 
+## 0.6.36
+
+### Patch Changes
+
+- a9471d0: fix sonar issues
+- Updated dependencies [a9471d0]
+    - @sap-ux/project-access@1.32.17
+    - @sap-ux/ui5-config@0.29.12
+
 ## 0.6.35
 
 ### Patch Changes